+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/_params.py
@@ -97,8 +97,6 @@ register_cli_argument('webapp', 'name', configured_default='web',
                       arg_type=name_arg_type, completer=get_resource_name_completion_list('Microsoft.Web/sites'), id_part='name',
                       help="name of the web. You can configure the default using 'az configure --defaults web=<name>'")
 register_cli_argument('webapp create', 'name', options_list=('--name', '-n'), help='name of the new webapp')
-register_cli_argument('webapp create', 'deployment_container_image_name', options_list=('--deployment-container-image-name', '-i'),
-                      help='Linux only. Container image name from Docker Hub, e.g. publisher/image-name:tag')
 register_cli_argument('webapp create', 'startup_file', help="Linux only. The web's startup file")
 register_cli_argument('webapp create', 'runtime', options_list=('--runtime', '-r'), help="canonicalized web runtime in the format of Framework|Version, e.g. \"PHP|5.6\". Use 'az webapp list-runtimes' for available list")  # TODO ADD completer
 register_cli_argument('webapp list-runtimes', 'linux', action='store_true', help='list runtime stacks for linux based webapps')  # TODO ADD completer
@@ -114,6 +112,7 @@ register_cli_argument('webapp delete', 'keep_metrics', action='store_true', help
 register_cli_argument('webapp delete', 'keep_dns_registration', action='store_true', help='keep DNS registration')
 
 for scope in ['webapp', 'functionapp']:
+    register_cli_argument(scope + ' create', 'deployment_container_image_name', options_list=('--deployment-container-image-name', '-i'), help='Linux only. Container image name from Docker Hub, e.g. publisher/image-name:tag')
     register_cli_argument(scope + ' config ssl bind', 'ssl_type', help='The ssl cert type', **enum_choice_list(['SNI', 'IP']))
     register_cli_argument(scope + ' config ssl upload', 'certificate_password', help='The ssl cert password')
     register_cli_argument(scope + ' config ssl upload', 'certificate_file', type=file_type, help='The filepath for the .pfx file')