+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/custom.py
@@ -9,8 +9,9 @@ try:
     from urllib.parse import urlparse
 except ImportError:
     from urlparse import urlparse  # pylint: disable=import-error
+from binascii import hexlify
+from os import urandom
 import OpenSSL.crypto
-
 from msrestazure.azure_exceptions import CloudError
 from msrestazure.tools import is_valid_resource_id, parse_resource_id
 
@@ -116,7 +117,7 @@ def list_webapp(resource_group_name=None):
 
 
 def list_function_app(resource_group_name=None):
-    return _list_app(['functionapp'], resource_group_name)
+    return _list_app(['functionapp', 'functionapp,linux'], resource_group_name)
 
 
 def _list_app(app_types, resource_group_name=None):
@@ -760,7 +761,7 @@ def show_backup_configuration(resource_group_name, webapp_name, slot=None):
     try:
         return _generic_site_operation(resource_group_name, webapp_name,
                                        'get_backup_configuration', slot)
-    except:
+    except Exception:  # pylint: disable=broad-except
         raise CLIError('Backup configuration not found')
 
 
@@ -1466,7 +1467,8 @@ class _StackRuntimeHelper(object):
 
 def create_function(resource_group_name, name, storage_account, plan=None,
                     consumption_plan_location=None, deployment_source_url=None,
-                    deployment_source_branch='master', deployment_local_git=None):
+                    deployment_source_branch='master', deployment_local_git=None,
+                    deployment_container_image_name=None):
     if deployment_source_url and deployment_local_git:
         raise CLIError('usage error: --deployment-source-url <url> | --deployment-local-git')
     if bool(plan) == bool(consumption_plan_location):
@@ -1475,28 +1477,46 @@ def create_function(resource_group_name, name, storage_account, plan=None,
     site_config = SiteConfig(app_settings=[])
     functionapp_def = Site(location=None, site_config=site_config)
     client = web_client_factory()
+
     if consumption_plan_location:
         locations = list_consumption_locations()
         location = next((l for l in locations if l['name'].lower() == consumption_plan_location.lower()), None)
         if location is None:
             raise CLIError("Location is invalid. Use: az functionapp list-consumption-locations")
         functionapp_def.location = consumption_plan_location
+        functionapp_def.kind = 'functionapp'
     else:
         if is_valid_resource_id(plan):
             plan = parse_resource_id(plan)['name']
         plan_info = client.app_service_plans.get(resource_group_name, plan)
         location = plan_info.location
+        is_linux = plan_info.reserved
+        if is_linux:
+            functionapp_def.kind = 'functionapp,linux'
+            site_config.app_settings.append(NameValuePair('FUNCTIONS_EXTENSION_VERSION', 'beta'))
+            site_config.app_settings.append(NameValuePair('MACHINEKEY_DecryptionKey',
+                                                          str(hexlify(urandom(32)).decode()).upper()))
+            if deployment_container_image_name:
+                site_config.app_settings.append(NameValuePair('DOCKER_CUSTOM_IMAGE_NAME',
+                                                              deployment_container_image_name))
+                site_config.app_settings.append(NameValuePair('FUNCTION_APP_EDIT_MODE', 'readOnly'))
+                site_config.app_settings.append(NameValuePair('WEBSITES_ENABLE_APP_SERVICE_STORAGE', 'false'))
+            else:
+                site_config.app_settings.append(NameValuePair('WEBSITES_ENABLE_APP_SERVICE_STORAGE', 'true'))
+                site_config.linux_fx_version = 'DOCKER|appsvc/azure-functions-runtime'
+        else:
+            functionapp_def.kind = 'functionapp'
+            site_config.app_settings.append(NameValuePair('FUNCTIONS_EXTENSION_VERSION', '~1'))
+
         functionapp_def.server_farm_id = plan
         functionapp_def.location = location
 
     con_string = _validate_and_get_connection_string(resource_group_name, storage_account)
-    functionapp_def.kind = 'functionapp'
 
     # adding appsetting to site to make it a function
     site_config.app_settings.append(NameValuePair('AzureWebJobsStorage', con_string))
     site_config.app_settings.append(NameValuePair('AzureWebJobsDashboard', con_string))
     site_config.app_settings.append(NameValuePair('WEBSITE_NODE_DEFAULT_VERSION', '6.5.0'))
-    site_config.app_settings.append(NameValuePair('FUNCTIONS_EXTENSION_VERSION', '~1'))
 
     if consumption_plan_location is None:
         site_config.always_on = True